net, dhcp: Move IPv6 server to serverv6 package
