+++ pkg/network/dhcp/serverv6/conn.go
@@ -16,7 +16,7 @@
  * Copyright 2020 Red Hat, Inc.
  *
  */
-package dhcpv6
+package serverv6
 
 import (
 	"fmt"
