+++ pkg/network/dhcp/serverv6/serverv6_suite_test.go
@@ -1,0 +1,11 @@
+package serverv6_test
+
+import (
+	"testing"
+
+	"kubevirt.io/client-go/testutils"
+)
+
+func TestServerv6(t *testing.T) {
+	testutils.KubeVirtTestSuiteSetup(t)
+}
