+++ pkg/network/dhcp/serverv6/serverv6_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package dhcpv6
+package serverv6
 
 import (
 	"net"
